[CSL-1621] Move Rubbish plugin into separate module
